+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/java/com/azure/android/communication/ui/callingcompositedemoapp/CallLauncherActivity.kt
@@ -9,8 +9,8 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.azure.android.communication.ui.callingcompositedemoapp.databinding.ActivityCallLauncherBinding
-import com.azure.android.communication.ui.callingcompositedemoapp.diagnostics.diagnosticsFeature
-import com.azure.android.communication.ui.callingcompositedemoapp.diagnostics.initializeMemoryViewFeature
+import com.azure.android.communication.ui.callingcompositedemoapp.features.AdditionalFeatures
+import com.azure.android.communication.ui.callingcompositedemoapp.features.conditionallyRegisterDiagnostics
 import com.azure.android.communication.ui.callingcompositedemoapp.launcher.CallingCompositeLauncher
 import com.azure.android.communication.ui.utilities.FeatureFlags
 import java.util.UUID
@@ -24,7 +24,9 @@ class CallLauncherActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         // Register Memory Viewer with FeatureFlags
-        initializeMemoryViewFeature(this)
+        conditionallyRegisterDiagnostics(this)
+        FeatureFlags.registerAdditionalFeature(AdditionalFeatures.secondaryThemeFeature)
+
         // Initialize the FeatureFlags enum
         FeatureFlags.initialize(this)
 
@@ -135,7 +137,7 @@ class CallLauncherActivity : AppCompatActivity() {
         super.onResume()
         // Will resume this feature if it's active
         // This is a special case for the permission dialog for diagnostic feature
-        diagnosticsFeature.onStart(application)
+        AdditionalFeatures.diagnosticsFeature.onStart(application)
     }
 
     override fun onDestroy() {
