+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/java/com/azure/android/communication/ui/callingcompositedemoapp/launcher/CallingCompositeKotlinLauncher.kt
@@ -3,6 +3,7 @@
 
 package com.azure.android.communication.ui.callingcompositedemoapp.launcher
 
+import com.azure.android.communication.calling.Features
 import com.azure.android.communication.common.CommunicationTokenCredential
 import com.azure.android.communication.common.CommunicationTokenRefreshOptions
 import com.azure.android.communication.ui.CallComposite
@@ -11,6 +12,9 @@ import com.azure.android.communication.ui.GroupCallOptions
 import com.azure.android.communication.ui.TeamsMeetingOptions
 import com.azure.android.communication.ui.callingcompositedemoapp.CallLauncherActivity
 import com.azure.android.communication.ui.callingcompositedemoapp.CallLauncherActivityErrorHandler
+import com.azure.android.communication.ui.callingcompositedemoapp.R
+import com.azure.android.communication.ui.callingcompositedemoapp.features.AdditionalFeatures
+import com.azure.android.communication.ui.configuration.ThemeConfiguration
 import java.util.UUID
 import java.util.concurrent.Callable
 
@@ -24,7 +28,11 @@ class CallingCompositeKotlinLauncher(private val tokenRefresher: Callable<String
         meetingLink: String?,
         showAlert: ((String) -> Unit)?,
     ) {
-        val callComposite: CallComposite = CallCompositeBuilder().build()
+        val callComposite: CallComposite =
+            if (AdditionalFeatures.secondaryThemeFeature.active)
+                CallCompositeBuilder().theme(ThemeConfiguration(R.style.MyCompany_Theme)).build()
+            else
+                CallCompositeBuilder().build()
 
         callComposite.setOnErrorHandler(CallLauncherActivityErrorHandler(callLauncherActivity))
