+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/res/values/strings.xml
@@ -18,4 +18,5 @@
     <string name="tokenFunctionRadioButtonText">Token Function</string>
     <string name="acsTokenRadioButtonText">@string/acsTokenHint</string>
     <string name="diagnostics">FPS, Memory Diagnostics</string>
+    <string name="secondary_theme">Secondary theme</string>
 </resources>
