+++ azure-communication-ui/azure-communication-ui-demo-app/src/main/res/values/themes.xml
@@ -20,8 +20,9 @@
     </style>
 
     <style name="MyCompany.Theme" parent="AzureCommunicationUI.Theme.Calling">
+        <item name="colorPrimary">@color/teal_700</item>
         <!-- Teams brand color. -->
-        <item name="azure_communication_ui_calling_primary_color">@color/Teams_light</item>
+        <item name="azure_communication_ui_calling_primary_color">?attr/colorPrimary</item>
     </style>
 
 </resources>
